Renamed Sofast.py to ProcesssofastFringe.py
